libffi: build DLL on Windows
The build setup is rather horrible, with some minimal gcc MSVC
wrapper. But the DLL is a prerequisite for the Python 3.8 build,
which dropped the internal libffi.

There is supposed to be a separate static Windows Arm64 build in
the msvc_build/aarch64 folder, but the DLL builds too, so this
just uses it for the moment.

This also defaults to explicit --buid and --host settings, even
without a cross build, because the prediced name would otherwise
differ (*-unknown-cygwin instead of *-pc-cygwin).

Additionally a "make install" also fails...

Change-Id: Ifb7dac840e23efffb9a5e342560aef9e11e0db79